+++ stories/table.stories.js
@@ -145,10 +145,6 @@ export const disableFilters = () => (
   />
 )
 
-export const title = () => (
-  <Table data={provinces} title='Cases by province' />
-)
-
 export const initialSortBy = () => (
   <Table
     data={provinces}